still making a lot of changes in code org, adding and moving tests

+++ hnsw/src/params.rs
@@ -1,30 +1,30 @@
 #[derive(Debug, Clone)]
 pub struct Params {
-    pub m: u8,
-    pub mmax: u8,
-    pub mmax0: u8,
+    pub m: usize,
+    pub mmax: usize,
+    pub mmax0: usize,
     pub ml: f32,
-    pub ef_cons: u32,
-    pub dim: u32,
+    pub ef_cons: usize,
+    pub dim: usize,
 }
 
-pub fn get_default_ml(m: u8) -> f32 {
+pub fn get_default_ml(m: usize) -> f32 {
     1.0 / (m as f32).ln()
 }
 
 impl Params {
-    pub fn from_m(m: u8, dim: u32) -> Params {
+    pub fn from_m(m: usize, dim: usize) -> Params {
         Params {
             m,
             mmax: m,
-            mmax0: (m * 2) as u8,
+            mmax0: m * 2,
             ml: get_default_ml(m),
-            ef_cons: (m as u32) * 2,
+            ef_cons: m * 2,
             dim,
         }
     }
 
-    pub fn from_m_efcons(m: u8, ef_cons: u32, dim: u32) -> Params {
+    pub fn from_m_efcons(m: usize, ef_cons: usize, dim: usize) -> Params {
         Params {
             m,
             mmax: m,
@@ -36,19 +36,19 @@ impl Params {
     }
 
     pub fn from(
-        m: u8,
-        ef_cons: Option<u32>,
-        mmax: Option<u8>,
-        mmax0: Option<u8>,
+        m: usize,
+        ef_cons: Option<usize>,
+        mmax: Option<usize>,
+        mmax0: Option<usize>,
         ml: Option<f32>,
-        dim: u32,
+        dim: usize,
     ) -> Params {
         Params {
             m,
             mmax: mmax.unwrap_or(m),
             mmax0: mmax0.unwrap_or(m * 2),
             ml: ml.unwrap_or(1.0 / (m as f32).ln()),
-            ef_cons: ef_cons.unwrap_or((m as u32) * 2),
+            ef_cons: ef_cons.unwrap_or(m * 2),
             dim,
         }
     }
